Support casting ip
