+++ core/src/main/java/org/opensearch/sql/calcite/ExtendedRexBuilder.java
@@ -6,13 +6,14 @@
 package org.opensearch.sql.calcite;
 
 import static org.opensearch.sql.calcite.utils.OpenSearchTypeFactory.ExprUDT.EXPR_DATE;
+import static org.opensearch.sql.calcite.utils.OpenSearchTypeFactory.ExprUDT.EXPR_IP;
 import static org.opensearch.sql.calcite.utils.OpenSearchTypeFactory.ExprUDT.EXPR_TIME;
 import static org.opensearch.sql.calcite.utils.OpenSearchTypeFactory.ExprUDT.EXPR_TIMESTAMP;
 
 import com.google.common.collect.ImmutableList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Set;
+import java.util.Locale;
 import org.apache.calcite.avatica.util.TimeUnit;
 import org.apache.calcite.rel.type.RelDataType;
 import org.apache.calcite.rex.RexBuilder;
@@ -24,7 +25,9 @@ import org.apache.calcite.sql.parser.SqlParserPos;
 import org.apache.calcite.sql.type.SqlTypeName;
 import org.apache.calcite.sql.type.SqlTypeUtil;
 import org.opensearch.sql.ast.expression.SpanUnit;
-import org.opensearch.sql.calcite.type.ExprSqlType;
+import org.opensearch.sql.calcite.type.AbstractExprRelDataType;
+import org.opensearch.sql.calcite.utils.OpenSearchTypeFactory;
+import org.opensearch.sql.exception.SemanticCheckException;
 import org.opensearch.sql.expression.function.PPLBuiltinOperators;
 
 public class ExtendedRexBuilder extends RexBuilder {
@@ -124,16 +127,16 @@ public class ExtendedRexBuilder extends RexBuilder {
         //            SqlStdOperatorTable.NOT_EQUALS,
         //            ImmutableList.of(exp, makeZeroLiteral(exp.getType())));
       }
-    } else if (type instanceof ExprSqlType exprSqlType
-        && Set.of(EXPR_DATE, EXPR_TIME, EXPR_TIMESTAMP).contains(exprSqlType.getUdt())) {
-      switch (exprSqlType.getUdt()) {
-        case EXPR_DATE:
-          return makeCall(type, PPLBuiltinOperators.DATE, List.of(exp));
-        case EXPR_TIME:
-          return makeCall(type, PPLBuiltinOperators.TIME, List.of(exp));
-        case EXPR_TIMESTAMP:
-          return makeCall(type, PPLBuiltinOperators.TIMESTAMP, List.of(exp));
-      }
+    } else if (OpenSearchTypeFactory.isUserDefinedType(type)) {
+      var udt = ((AbstractExprRelDataType<?>) type).getUdt();
+      return switch (udt) {
+        case EXPR_DATE -> makeCall(type, PPLBuiltinOperators.DATE, List.of(exp));
+        case EXPR_TIME -> makeCall(type, PPLBuiltinOperators.TIME, List.of(exp));
+        case EXPR_TIMESTAMP -> makeCall(type, PPLBuiltinOperators.TIMESTAMP, List.of(exp));
+        case EXPR_IP -> makeCall(type, PPLBuiltinOperators.CAST_IP, List.of(exp));
+        default -> throw new SemanticCheckException(
+            String.format(Locale.ROOT, "Unsupported cast type: %s", udt.name()));
+      };
     }
     return super.makeCast(pos, type, exp, matchNullability, safe, format);
   }
