+++ core/src/main/java/org/opensearch/sql/calcite/utils/OpenSearchTypeFactory.java
@@ -259,7 +259,8 @@ public class OpenSearchTypeFactory extends JavaTypeFactoryImpl {
 
   /** Converts a Calcite data type to OpenSearch ExprCoreType. */
   public static ExprType convertRelDataTypeToExprType(RelDataType type) {
-    if (type instanceof AbstractExprRelDataType<?> udt) {
+    if (isUserDefinedType(type)) {
+      AbstractExprRelDataType<?> udt = (AbstractExprRelDataType<?>) type;
       return udt.getExprType();
     }
     ExprType exprType = convertSqlTypeNameToExprType(type.getSqlTypeName());
@@ -325,5 +326,15 @@ public class OpenSearchTypeFactory extends JavaTypeFactoryImpl {
       return exprRelDataType.getJavaType();
     }
     return super.getJavaClass(type);
+  }
+
+  /**
+   * Whether a given RelDataType is a user-defined type (UDT)
+   *
+   * @param type the RelDataType to check
+   * @return true if the type is a user-defined type, false otherwise
+   */
+  public static boolean isUserDefinedType(RelDataType type) {
+    return type instanceof AbstractExprRelDataType<?>;
   }
 }
