+++ core/src/main/java/org/opensearch/sql/expression/function/PPLBuiltinOperators.java
@@ -75,6 +75,7 @@ import org.opensearch.sql.expression.function.udf.datetime.WeekdayFunction;
 import org.opensearch.sql.expression.function.udf.datetime.YearweekFunction;
 import org.opensearch.sql.expression.function.udf.ip.CidrMatchFunction;
 import org.opensearch.sql.expression.function.udf.ip.CompareIpFunction;
+import org.opensearch.sql.expression.function.udf.ip.IPCastFunction;
 import org.opensearch.sql.expression.function.udf.math.CRC32Function;
 import org.opensearch.sql.expression.function.udf.math.ConvFunction;
 import org.opensearch.sql.expression.function.udf.math.DivideFunction;
@@ -281,6 +282,9 @@ public class PPLBuiltinOperators extends ReflectiveSqlOperatorTable {
               NullPolicy.ARG0,
               PPLOperandTypes.DATETIME_OR_STRING)
           .toUDF("TIME");
+
+  // IP cast function
+  public static final SqlOperator CAST_IP = new IPCastFunction().toUDF("CAST_IP");
   public static final SqlOperator TIME_TO_SEC =
       adaptExprMethodToUDF(
               DateTimeFunctions.class,
