+++ core/src/main/java/org/opensearch/sql/expression/function/UDFOperandMetadata.java
@@ -176,4 +176,40 @@ public interface UDFOperandMetadata extends SqlOperandMetadata {
       return "";
     }
   }
+
+  /**
+   * A named class that serves as an identifier for IP cast's operand metadata. It does not
+   * implement any actual type checking logic.
+   */
+  class IPCastOperandMetadata implements UDFOperandMetadata {
+    @Override
+    public SqlOperandTypeChecker getInnerTypeChecker() {
+      return this;
+    }
+
+    @Override
+    public List<RelDataType> paramTypes(RelDataTypeFactory typeFactory) {
+      return List.of();
+    }
+
+    @Override
+    public List<String> paramNames() {
+      return List.of();
+    }
+
+    @Override
+    public boolean checkOperandTypes(SqlCallBinding callBinding, boolean throwOnFailure) {
+      return false;
+    }
+
+    @Override
+    public SqlOperandCountRange getOperandCountRange() {
+      return null;
+    }
+
+    @Override
+    public String getAllowedSignatures(SqlOperator op, String opName) {
+      return "";
+    }
+  }
 }
